Minimal fix for disappearing indicator
An odd value of fe.height leads to the indicator disappearing. It seems
that buffer_release is not called through the listener in this case.
As a result buffer->busy stays true.

This commit provides the minimal changes required to resolve this
behavior. They all interact with the wayland shm buffer protocol.